rename test_fixture to testbench
